refactor(core): split pipeline types and color scale resolution
Separate public contract types from LayerFrame/binding types behind a stable
types barrel. Split sequential vs ordinal color resolution out of
resolveColorScale. Characterization tests pin sequential inference, explicit
ordinal override, and colorOf unknown fallback.
